RAG_local/hyde/requirements.txt
# requirements.txt
openai>=0.27.0
pinecone-client>=2.2.1
python-dotenv>=0.19.0
numpy>=1.21.0